Refactor: extract bindings and graph renderers into modules
