+++ crates/solidity/outputs/cargo/tests/src/bindings_output/mod.rs
@@ -1,2 +1,4 @@
 mod generated;
+mod graph;
+mod renderer;
 mod runner;
